LayoutTests/http/tests/inspector/filesystem/request-metadata-expected.txt

Tests requestMetadata command.

errorCode: 2
metadata: (null)

errorCode: 8
metadata: (null)

errorCode: 0
metadata:
  modificationTime: (exists)
  size: 4

errorCode: 0
metadata:
  modificationTime: (exists)
  size: 0

errorCode: 1
metadata: (null)

All tests have done.
